+++ lydian/bot.py
@@ -1,5 +1,6 @@
 """Handles bot setup and execution."""
 import asyncio
+import inspect
 import logging
 import os
 import sys
@@ -8,7 +9,8 @@ from datetime import UTC, datetime
 from getpass import getpass
 
 import discord
-from discord.ext import commands
+from benedict import benedict
+from discord.ext import commands, tasks
 from dotenv import load_dotenv
 from loguru import logger
 from rich.prompt import Confirm
@@ -17,7 +19,6 @@ from lydian.cogs.debug import DebugCog
 from lydian.cogs.general import GeneralCog
 from lydian.cogs.util import embed_error, embed_info
 from lydian.cogs.voice import VoiceCog
-from lydian.cogs.voice import background_tasks as voice_background_tasks
 from lydian.config import config
 from lydian.console import bot_console
 from lydian.const import (
@@ -28,6 +29,7 @@ from lydian.const import (
     LOGS_DIR,
     PERMISSIONS_PATH,
     PROJECT_VERSION,
+    LogLevel,
     clear_tmp_dir,
     create_directories,
     debug_context,
@@ -36,9 +38,41 @@ from lydian.const import (
 )
 from lydian.errors import AbortCommand
 from lydian.perms import PERMISSIONS_DEFAULT, perms
-from lydian.util import dirsize
+from lydian.util import dirsize, get_background_tasks, get_leaves
 
 load_dotenv('.env')
+
+class InterceptHandler(logging.Handler):
+    """Handles redirecting logs for the ``discord`` module to ``loguru``."""
+
+    def emit(self, record: logging.LogRecord) -> None:  # noqa: D102
+        # Get corresponding Loguru level if it exists.
+        try:
+            level: str | int = logger.level(record.levelname).name
+        except ValueError:
+            level = record.levelno
+
+        if record.funcName.startswith('_'):
+            level = LogLevel.DEBUG
+
+        # Find caller from where originated the logged message.
+        frame, depth = inspect.currentframe(), 0
+        while frame:
+            filename = frame.f_code.co_filename
+            is_logging = filename == logging.__file__
+            is_frozen = 'importlib' in filename and '_bootstrap' in filename
+            if depth > 0 and not (is_logging or is_frozen):
+                break
+            frame = frame.f_back
+            depth += 1
+
+        logger.opt(depth=depth, exception=record.exc_info).log(level, record.getMessage())
+
+# Redirect discord.py logs to our logger
+discord_logger = logging.getLogger('discord')
+discord_logger.setLevel(config.logging.log_level)
+logging.getLogger('discord.http').setLevel(config.logging.log_level)
+discord_logger.addHandler(InterceptHandler())
 
 intents = discord.Intents.default()
 intents.messages = True
@@ -48,12 +82,8 @@ intents.voice_states = True
 bot = commands.Bot(
     intents=intents,
     command_prefix=config.prefix,
-    log_handler=logging.FileHandler(filename=LOGS_DIR / 'discord.log', encoding='utf-8', mode='w'),
 )
-
 debug_context['bot'] = bot
-
-event_start_console = asyncio.Event()
 
 @bot.event
 async def on_message(message: discord.Message) -> None:
@@ -104,11 +134,13 @@ async def thread_bot() -> None:
     async with bot:
         # Add cogs
         await bot.add_cog(GeneralCog(bot))
-        await bot.add_cog(cog_voice := VoiceCog(bot))
+        await bot.add_cog(VoiceCog(bot))
         if config.debug:
             await bot.add_cog(DebugCog(bot))
 
-        debug_context['cog.voice'] = cog_voice
+        debug_context['cog'] = {name.removesuffix('Cog').lower():cog for name, cog in bot.cogs.items()}
+        debug_context['tasks'] = benedict(get_background_tasks(bot))
+        debug_context['tasklist'] = list(get_leaves(debug_context['tasks'], tasks.Loop))
 
         # Start
         logger.info('Logging in; wait for "Ready!" before running commands')
@@ -118,16 +150,26 @@ async def thread_bot() -> None:
             logger.error(f'No bot token found, please set the {target_env_var} environment variable')
             return
 
-        event_start_console.set()
         debug_context['bot-start-time'] = datetime.now(UTC)
-        await bot.start(token)
+        try:
+            await bot.start(token)
+        except discord.LoginFailure as e:
+            if 'Improper token' in str(e):
+                logger.error('Bad token given; check your'
+                    + f' {'LYDIAN_DEBUG_TOKEN' if config.debug else 'LYDIAN_TOKEN'} value')
+            else:
+                logger.opt(exception=e).error(f'Failed to log in: {e}')
 
 async def thread_console() -> None:
     """Returns the ``Coroutine`` thread for the interactive console."""
-    await event_start_console.wait()
+    while not bot.user:  # noqa: ASYNC110
+        # For ignoring ASYNC110: Login failures cause issues with a running console, so we need to make sure the bot is
+        # logged in first, and since `await bot.start()` blocks until's connection is closed, we have no way of ensuring
+        # that through an event alone. This is the only way to do this for now.
+        await asyncio.sleep(1)
 
     bot_console.bot = bot
-    await bot_console.start_loop()
+    await bot_console.start_loop(catch=True)
 
 def prompt_bot_setup() -> bool:
     """Prompts the user to setup Lydian in the current working directory and returns whether the user confirmed."""
@@ -187,7 +229,9 @@ async def async_main() -> int:
         return_when=asyncio.FIRST_COMPLETED,
     )
 
-    for task in voice_background_tasks:
+    tasklist: list[tasks.Loop] = debug_context['tasklist']
+
+    for task in tasklist:
         if (internal := task.get_task()) and internal.done():
             _ = internal.exception()