+++ tests/test_util.py
@@ -138,9 +138,57 @@ def test_get_dataclass_fields() -> None:
     assert dc_fields['d.x'].default == dc.d.x
     assert dc_fields['d.y'].default_factory() == dc.d.y  # ty:ignore[call-non-callable]
 
+def test_get_leaves() -> None:
+    tree = {
+        'a': 1,
+        'b': '2',
+        'c': {
+            'a': 3,
+            'b': {
+                'a': 4,
+            },
+            'c': '5',
+        },
+        'd': '6',
+    }
+
+    assert list(util.get_leaves(tree)) == [1, '2', 3, 4, '5', '6']
+    assert list(util.get_leaves(tree, int)) == [1, 3, 4]
+    assert list(util.get_leaves(tree, str)) == ['2', '5', '6']
+
 def test_is_annotated() -> None:
     assert not util.is_annotated(str)
     assert util.is_annotated(Annotated[str, 'Description'])
+
+def test_iter_columns() -> None:
+    table = [
+        [1, 2, 3],
+        [4, 5, 6],
+        [7, 8, 9],
+    ]
+
+    assert list(util.iter_columns(table)) == [(1, 4, 7), (2, 5, 8), (3, 6, 9)]
+
+    uneven_table = [
+        [1,  2,  3],
+        [4,  5,  6,  7, 8],
+        [9],
+        [10, 11, 12, 13],
+        [14, 15, 16, 17],
+        [],
+        [18, 19],
+    ]
+
+    with pytest.raises(IndexError):
+        assert list(util.iter_columns(uneven_table))
+
+    assert list(util.iter_columns(uneven_table, 0)) == [
+        (1, 4, 9, 10, 14, 0, 18),
+        (2, 5, 0, 11, 15, 0, 19),
+        (3, 6, 0, 12, 16, 0, 0),
+        (0, 7, 0, 13, 17, 0, 0),
+        (0, 8, 0,  0,  0, 0, 0),
+    ]
 
 def test_join_trailing() -> None:
     assert util.join_trailing('abc', ' ') == 'a b c '